TestingNewCode.py - Addition - New Main Py file
This is the new main .py file. This file allows us to have all the motors move in unison...hopefully we can add tracing and maybe, just maybe, we can move multiple motors at the same time. This should be able control all the motors and the sensors at the same time...no promises.